🎨 Moved app JSON encoder to utils
